Add baseline profile modules per library. The generateBaselineProfile gradle task doesn't seem to support passing an argument for which test to run so it would run each test per app/library.
Change demo app package structure to disambiguate library and demo app code.
