--- app/src/main/kotlin/com/alexrdclement/uiplayground/components/Component.kt
@@ -1,9 +1,0 @@
-package com.alexrdclement.uiplayground.components
-
-import com.alexrdclement.uiplayground.catalog.CatalogItem
-
-enum class Component : CatalogItem {
-    MediaControlSheet;
-
-    override val title = this.name
-}
