--- app/src/main/kotlin/com/alexrdclement/uiplayground/components/ComponentScreen.kt
@@ -1,13 +1,0 @@
-package com.alexrdclement.uiplayground.components
-
-import androidx.compose.runtime.Composable
-import com.alexrdclement.uiplayground.components.demo.MediaControlSheetDemo
-
-@Composable
-fun ComponentScreen(
-    component: Component,
-) {
-    when (component) {
-        Component.MediaControlSheet -> MediaControlSheetDemo()
-    }
-}
