+++ app/src/main/kotlin/com/alexrdclement/uiplayground/demo/components/Component.kt
@@ -1,0 +1,9 @@
+package com.alexrdclement.uiplayground.demo.components
+
+import com.alexrdclement.uiplayground.catalog.CatalogItem
+
+enum class Component : CatalogItem {
+    MediaControlSheet;
+
+    override val title = this.name
+}
