+++ app/src/main/kotlin/com/alexrdclement/uiplayground/demo/components/ComponentScreen.kt
@@ -1,0 +1,13 @@
+package com.alexrdclement.uiplayground.demo.components
+
+import androidx.compose.runtime.Composable
+import com.alexrdclement.uiplayground.demo.components.demo.MediaControlSheetDemo
+
+@Composable
+fun ComponentScreen(
+    component: Component,
+) {
+    when (component) {
+        Component.MediaControlSheet -> MediaControlSheetDemo()
+    }
+}
