--- app/src/main/kotlin/com/alexrdclement/uiplayground/shaders/navigation/ShadersNavigation.kt
@@ -1,20 +1,0 @@
-package com.alexrdclement.uiplayground.shaders.navigation
-
-import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.compose.composable
-import com.alexrdclement.uiplayground.shaders.ShaderScreen
-
-private const val shadersRoute = "shaders"
-
-fun NavController.navigateToShaders() {
-    this.navigate(route = shadersRoute) {
-        launchSingleTop = true
-    }
-}
-
-fun NavGraphBuilder.shadersScreen() {
-    composable(shadersRoute) {
-        ShaderScreen()
-    }
-}
