+++ baseline-profile/src/main/AndroidManifest.xml
