+++ baseline-profile/src/main/kotlin/com/alexrdclement/uiplayground/baselineprofile/BaselineProfileGenerator.kt
@@ -1,0 +1,29 @@
+package com.alexrdclement.uiplayground.baselineprofile
+
+import androidx.benchmark.macro.junit4.BaselineProfileRule
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.filters.LargeTest
+import com.alexrdclement.uiplayground.appPackageName
+import org.junit.Rule
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+@LargeTest
+class BaselineProfileGenerator {
+
+    @get:Rule
+    val rule = BaselineProfileRule()
+
+    @Test
+    fun generateAppProfile() {
+        rule.collect(
+            packageName = appPackageName,
+            // See: https://d.android.com/topic/performance/baselineprofiles/dex-layout-optimizations
+            includeInStartupProfile = true
+        ) {
+            pressHome()
+            startActivityAndWait()
+        }
+    }
+}
