--- baselineprofile/src/main/kotlin/com/alexrdclement/uiplayground/baselineprofile/BaselineProfileGenerator.kt
@@ -1,29 +1,0 @@
-package com.alexrdclement.uiplayground.baselineprofile
-
-import androidx.benchmark.macro.junit4.BaselineProfileRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.filters.LargeTest
-import com.alexrdclement.uiplayground.packageName
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@RunWith(AndroidJUnit4::class)
-@LargeTest
-class BaselineProfileGenerator {
-
-    @get:Rule
-    val rule = BaselineProfileRule()
-
-    @Test
-    fun generate() {
-        rule.collect(
-            packageName = packageName,
-            // See: https://d.android.com/topic/performance/baselineprofiles/dex-layout-optimizations
-            includeInStartupProfile = true
-        ) {
-            pressHome()
-            startActivityAndWait()
-        }
-    }
-}
