+++ benchmark/src/main/kotlin/com/alexrdclement/uiplayground/shaders/PixelateBenchmark.kt
@@ -1,6 +1,5 @@
 package com.alexrdclement.uiplayground.shaders
 
-import android.graphics.Point
 import androidx.benchmark.macro.ExperimentalMetricApi
 import androidx.benchmark.macro.FrameTimingMetric
 import androidx.benchmark.macro.StartupMode
@@ -8,10 +7,8 @@ import androidx.benchmark.macro.TraceSectionMetric
 import androidx.benchmark.macro.TraceSectionMetric.Mode
 import androidx.benchmark.macro.junit4.MacrobenchmarkRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.uiautomator.By
 import com.alexrdclement.uiplayground.MainCatalogPage
-import com.alexrdclement.uiplayground.packageName
-import com.alexrdclement.uiplayground.waitAndFindObject
+import com.alexrdclement.uiplayground.appPackageName
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -21,10 +18,12 @@ class PixelateBenchmark {
     @get:Rule
     val benchmarkRule = MacrobenchmarkRule()
 
+    lateinit var shadersPage: ShadersPage
+
     @OptIn(ExperimentalMetricApi::class)
     @Test
     fun amountAdjustment() = benchmarkRule.measureRepeated(
-        packageName = packageName,
+        packageName = appPackageName,
         metrics = listOf(
             FrameTimingMetric(),
             TraceSectionMetric("pixelate", Mode.Sum),
@@ -36,12 +35,9 @@ class PixelateBenchmark {
             startActivityAndWait()
 
             MainCatalogPage(device).navigateToShaders()
-            ShadersPage(device).selectModifier("Pixelate")
+            shadersPage = ShadersPage(device).apply { selectPixelate() }
         }
     ) {
-        // One adjustment doesn't generate frame data
-        val xAmount = device.waitAndFindObject(By.desc("Amount"))
-        xAmount.drag(Point(xAmount.visibleCenter.x + 100, xAmount.visibleCenter.y))
-        xAmount.drag(Point(xAmount.visibleCenter.x + 200, xAmount.visibleCenter.y))
+        shadersPage.adjustPixelate()
     }
 }
