+++ shaders/baseline-profile/build.gradle.kts
@@ -1,0 +1,43 @@
+plugins {
+    alias(libs.plugins.androidTest)
+    alias(libs.plugins.kotlinAndroid)
+    alias(libs.plugins.baselineprofile)
+}
+
+android {
+    namespace = "com.alexrdclement.uiplayground.shaders.baselineprofile"
+    compileSdk = 34
+
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_1_8
+        targetCompatibility = JavaVersion.VERSION_1_8
+    }
+
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+
+    defaultConfig {
+        minSdk = 28
+        targetSdk = 34
+
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+    }
+
+    targetProjectPath = ":app"
+}
+
+// This is the configuration block for the Baseline Profile plugin.
+// You can specify to run the generators on a managed devices or connected devices.
+baselineProfile {
+    useConnectedDevices = true
+}
+
+dependencies {
+    implementation(libs.androidx.test.ext.junit)
+    implementation(libs.espresso.core)
+    implementation(libs.androidx.uiautomator)
+    implementation(libs.androidx.benchmark.macro.junit4)
+
+    implementation(projects.uiautomatorFixtures)
+}
