+++ uiautomator-fixtures/src/main/kotlin/com/alexrdclement/uiplayground/Constants.kt
@@ -1,3 +1,5 @@
 package com.alexrdclement.uiplayground
 
-const val packageName = "com.alexrdclement.uiplayground"
+const val appPackageName = "com.alexrdclement.uiplayground"
+const val componentsPackageName = "com.alexrdclement.uiplayground.components"
+const val shadersPackageName = "com.alexrdclement.uiplayground.shaders"
