+++ uiautomator-fixtures/src/main/kotlin/com/alexrdclement/uiplayground/shaders/ShadersPage.kt
@@ -1,5 +1,6 @@
 package com.alexrdclement.uiplayground.shaders
 
+import android.graphics.Point
 import androidx.test.uiautomator.By
 import androidx.test.uiautomator.UiDevice
 import com.alexrdclement.uiplayground.waitAndFindObject
@@ -7,8 +8,42 @@ import com.alexrdclement.uiplayground.waitAndFindObject
 class ShadersPage(
     private val device: UiDevice,
 ) {
-    fun selectModifier(modifierName: String) {
-        device.waitAndFindObject(By.text("None")).click()
+
+    fun selectChromaticAberration() {
+        selectModifier("Chromatic Aberration")
+    }
+
+    fun adjustChromaticAberration() {
+        val xAmount = device.waitAndFindObject(By.desc("X Amount"))
+        xAmount.drag(Point(xAmount.visibleCenter.x + 100, xAmount.visibleCenter.y))
+        val yAmount = device.waitAndFindObject(By.desc("Y Amount"))
+        yAmount.drag(Point(yAmount.visibleCenter.x + 100, yAmount.visibleCenter.y))
+    }
+
+    fun selectNoise() {
+        selectModifier("Noise")
+    }
+
+    fun adjustNoise() {
+        // One adjustment doesn't generate benchmark/profile frame data
+        val xAmount = device.waitAndFindObject(By.desc("Amount"))
+        xAmount.drag(Point(xAmount.visibleCenter.x + 100, xAmount.visibleCenter.y))
+        xAmount.drag(Point(xAmount.visibleCenter.x + 200, xAmount.visibleCenter.y))
+    }
+
+    fun selectPixelate() {
+        selectModifier("Pixelate")
+    }
+
+    fun adjustPixelate() {
+        // One adjustment doesn't generate benchmark/profile frame data
+        val xAmount = device.waitAndFindObject(By.desc("Amount"))
+        xAmount.drag(Point(xAmount.visibleCenter.x + 100, xAmount.visibleCenter.y))
+        xAmount.drag(Point(xAmount.visibleCenter.x + 200, xAmount.visibleCenter.y))
+    }
+
+    private fun selectModifier(modifierName: String) {
+        device.waitAndFindObject(By.desc("Select modifier")).click()
         device.waitAndFindObject(By.text(modifierName)).click()
     }
 }
